pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finetree-annotator"
version = "0.1.0"
description = "PDF page annotation tools for FineTree"
requires-python = ">=3.12"
dependencies = [
  "pydantic==2.12.5",
  "pdf2image",
  "PyQt5>=5.15,<5.16",
  "google-genai",
  "PyYAML",
  "Pillow",
  "transformers",
  "datasets",
  "httpx",
  "openai",
  "scikit-learn",
  "trl",
  "peft",
  "fastapi",
  "uvicorn",
  "python-multipart",
  "huggingface_hub",
  "runpod",
  "qwen-vl-utils",
  "gradio",
  "pytest>=8,<9",
  "pytest-qt>=4.4,<5",
]

[project.optional-dependencies]
ocr = [
  "python-doctr",
]
train = [
  "unsloth; platform_system == 'Linux'",
]

[project.scripts]
finetree-annotator = "finetree_annotator.app:main"
finetreeannotator = "finetree_annotator.app:main"
finetree-pdf-to-images = "finetree_annotator.pdf_to_images:main"
finetree-gemini-vlm = "finetree_annotator.gemini_vlm:main"
finetree-qwen-gt = "finetree_annotator.qwen_vlm:main"
finetree-backup-data = "finetree_annotator.workspace_backup:main"
list_commands = "finetree_annotator.list_commands:main"
finetree-ft-build-dataset = "finetree_annotator.finetune.dataset_builder:main"
finetree-ft-preflight = "finetree_annotator.finetune.preflight:main"
finetree-ft-train = "finetree_annotator.finetune.trainer_unsloth:main"
finetree-ft-merge-push = "finetree_annotator.finetune.merge_push:main"
finetree-ft-export-quantized = "finetree_annotator.finetune.export_quantized:main"
finetree-ft-push-dataset = "finetree_annotator.finetune.push_dataset_hub:main"
finetree-ft-push-dataset-no-bbox = "finetree_annotator.finetune.push_dataset_hub_no_bbox:main"
finetree-runpod-endpoint-env = "finetree_annotator.deploy.runpod_endpoint:main"
finetree-runpod-worker = "finetree_annotator.deploy.runpod_serverless_worker:main"
finetree-pod-api = "finetree_annotator.deploy.pod_api:main"
finetree-simple-infer-api = "finetree_annotator.deploy.simple_infer_api:main"
finetree-pod-gradio = "finetree_annotator.deploy.pod_gradio:main"
finetree-runpod-pod-start = "finetree_annotator.deploy.runpod_pod_start:main"
benchmark-new = "benchmark_new.cli:main"
"benchmark_new.cli" = "benchmark_new.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
finetree_annotator = ["assets/icons/*.png"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["finetree_annotator*", "benchmark_new*"]
